+++ src/converters/piral-blazor/src/interop.ts
@@ -1,5 +1,5 @@
 import { PiletApi } from 'piral-core';
-import { emitRenderEvent, emitNavigateEvent, emitPiralEvent } from './events';
+import { emitRenderEvent, emitUpdateEvent, emitNavigateEvent, emitPiralEvent, blazorRootId } from './events';
 import type { BlazorLogLevel, BlazorRootConfig, WebAssemblyStartOptions } from './types';
 
 const wasmLib = 'Microsoft.AspNetCore.Components.WebAssembly';
@@ -36,7 +36,7 @@ function createBase() {
   return document.head.appendChild(el);
 }
 
-function prepareForStartup() {
+async function prepareForStartup() {
   const originalApplyHotReload = window.Blazor._internal.applyHotReload;
   const queue = [];
 
@@ -56,17 +56,23 @@ function prepareForStartup() {
     queue.push(() => originalApplyHotReload.apply(this, args));
   };
 
-  return getCapabilities().then((capabilities) => ({
+  const capabilities = await getCapabilities();
+
+  if (capabilities.includes('custom-element')) {
+    document.getElementById(blazorRootId).setAttribute('render', 'modern');
+  }
+
+  return {
     capabilities,
     applyChanges,
-  }));
+  };
 }
 
 function createBlazorStarter(publicPath: string): (opts: WebAssemblyStartOptions) => Promise<BlazorRootConfig> {
   const root = document.body.appendChild(document.createElement('div'));
 
   root.style.display = 'contents';
-  root.id = 'blazor-root';
+  root.id = blazorRootId;
 
   if (publicPath) {
     const baseElement = document.head.querySelector('base') || createBase();
@@ -261,6 +267,7 @@ export function initialize(scriptUrl: string, publicPath: string, opts: WebAssem
     script.onload = () => {
       Object.assign(window.Blazor, {
         emitRenderEvent,
+        emitUpdateEvent,
         emitNavigateEvent,
         emitPiralEvent,
       });
@@ -275,8 +282,10 @@ export function initialize(scriptUrl: string, publicPath: string, opts: WebAssem
 export function createBootLoader(scriptUrl: string, extraScriptUrls: Array<string>) {
   const publicPath = computePath();
 
-  return (opts?: WebAssemblyStartOptions) => {
-    if (typeof window.$blazorLoader === 'undefined') {
+  return async (opts?: WebAssemblyStartOptions) => {
+    const first = typeof window.$blazorLoader === 'undefined';
+
+    if (first) {
       window.dispatchEvent(new CustomEvent('loading-blazor-core'));
 
       // we load all satellite scripts before we initialize blazor
@@ -285,6 +294,10 @@ export function createBootLoader(scriptUrl: string, extraScriptUrls: Array<strin
       );
     }
 
-    return window.$blazorLoader;
+    const config = await window.$blazorLoader;
+    return {
+      config,
+      first,
+    };
   };
 }