+++ src/converters/piral-vue/convert.ts
@@ -1,4 +1,4 @@
-import { createConverter } from './esm/converter';
+import { createConverter } from './lib/converter';
 
 export interface HtmlComponent<TProps> {
   component: {